scripts/50/requirements.txt
moviepy==1.0.3
Pillow==9.2.0